+++ nvidia/NVIDIA.h
@@ -4,6 +4,10 @@
 #include <memory>
 #include <string_view>
 
+#if defined(PMT_CUDAWRAPPERS_API)
+#include <cudawrappers/cu.hpp>
+#endif
+
 #include "common/PMT.h"
 
 namespace pmt::nvidia {
@@ -11,6 +15,9 @@ class NVIDIA : public PMT {
  public:
   constexpr static inline std::string_view name = "nvidia";
   static std::unique_ptr<PMT> Create(int device_number = 0);
+#if defined(PMT_CUDAWRAPPERS_API)
+  static std::unique_ptr<PMT> Create(cu::Device& device);
+#endif
 };
 }  // end namespace pmt::nvidia